requirements.txt
elasticsearch==8.17.2
fastmcp==2.2.6
python-dotenv